test(select): migrate custom
