+++ core/src/components/select/test/custom/select.e2e.ts-snapshots/select-custon-diff-ios-ltr-Mobile-Firefox-linux.png
