+++ core/src/components/select/test/custom/select.e2e.ts-snapshots/select-custon-diff-md-ltr-Mobile-Chrome-linux.png
