+++ core/src/components/select/test/custom/select.e2e.ts-snapshots/select-custon-diff-md-ltr-Mobile-Safari-linux.png
